+++ integreat_cms/cms/models/pois/poi.py
@@ -2,10 +2,21 @@ from django.db import models
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.utils.translation import ugettext_lazy as _
 
+from ...utils.translation_utils import ugettext_many_lazy as __
 from ..abstract_content_model import AbstractContentModel
 from ..media.media_file import MediaFile
 from ..pois.poi_translation import POITranslation
 from ..poi_categories.poi_category import POICategory
+
+
+def get_default_opening_hours():
+    """
+    Return the default opening hours
+
+    :return: The default opening hours
+    :rtype: list
+    """
+    return [{"allDay": False, "closed": True, "timeSlots": []} for _ in range(7)]
 
 
 class POI(AbstractContentModel):
@@ -66,6 +77,18 @@ class POI(AbstractContentModel):
         on_delete=models.SET_NULL,
         related_name="pois",
         verbose_name=_("category"),
+    )
+    temporarily_closed = models.BooleanField(
+        default=False,
+        verbose_name=_("temporarily closed"),
+        help_text=__(
+            _("Whether or not the location is temporarily closed."),
+            _("The opening hours remain and are only hidden."),
+        ),
+    )
+    opening_hours = models.JSONField(
+        default=get_default_opening_hours,
+        verbose_name=_("opening hours"),
     )
 
     @property